+++ tests/integration/account-streamer.test.ts
@@ -1,18 +1,19 @@
 import { AccountStreamer } from "../../lib/account-streamer";
 import { TastytradeLogger } from "../../lib/logger";
 import TastytradeHttpClient from "../../lib/services/tastytrade-http-client";
-import SessionService from "../../lib/services/session-service";
 
-const client = new TastytradeHttpClient({ baseUrl: process.env.BASE_URL! })
-const accountStreamer = new AccountStreamer(process.env.STREAMER_URL!, client.session, client.accessToken, new TastytradeLogger())
-
-beforeAll(async () => {
-  const sessionService = new SessionService(client)
-  await sessionService.login(process.env.API_USERNAME!, process.env.API_PASSWORD!)
-});
+const client = new TastytradeHttpClient({
+  baseUrl: process.env.BASE_URL!,
+  clientSecret: process.env.OAUTH_CLIENT_SECRET!,
+  refreshToken: process.env.OAUTH_REFRESH_TOKEN!,
+  oauthScopes: ['read']
+})
+const accountStreamer = new AccountStreamer(process.env.STREAMER_URL!, client, new TastytradeLogger())
 
 describe('subscribeToAccounts', () => {
   it('subscribes successfully', async function() {
+    console.log(process.env.STREAMER_URL)
+    console.log(process.env.API_ACCOUNT_NUMBER)
     await accountStreamer.start()
     await accountStreamer.subscribeToAccounts([process.env.API_ACCOUNT_NUMBER!])
   })